chore: moved `internalLink` to `utils` folder
