+++ ts/hooks/useOpenDeepLink.ts
@@ -1,12 +1,7 @@
 import { useLinkTo } from "@react-navigation/native";
 import * as O from "fp-ts/lib/Option";
 import { pipe } from "fp-ts/lib/function";
-import {
-  IO_FIMS_LINK_PREFIX,
-  IO_INTERNAL_LINK_PREFIX,
-  IO_UNIVERSAL_LINK_PREFIX
-} from "../utils/navigation";
-import { extractPathFromURL } from "../utils/url";
+import { getInternalRoute } from "../utils/internalLink";
 
 /**
  * This hook handles deep links. It removes the prefix and navigates to the path using the linkTo function
@@ -16,17 +11,5 @@ export const useOpenDeepLink = () => {
   const linkTo = useLinkTo();
 
   return (url: string) =>
-    pipe(
-      extractPathFromURL(
-        [
-          IO_INTERNAL_LINK_PREFIX,
-          IO_UNIVERSAL_LINK_PREFIX,
-          IO_FIMS_LINK_PREFIX
-        ],
-        url
-      ),
-      O.fromNullable,
-      O.map(path => (path.startsWith("/") ? path : "/" + path)),
-      O.map(linkTo)
-    );
+    pipe(getInternalRoute(url), O.fromNullable, O.map(linkTo));
 };
